@sergeylunev 2022-06-26
До оптимизации
```
goos: linux
goarch: amd64
pkg: github.com/fixme_my_friend/hw10_program_optimization
cpu: Intel(R) Core(TM) i3-10100F CPU @ 3.60GHz
BenchmarkGetDomainStat-8   	       8	 126403968 ns/op	134445110 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-8   	       8	 126629685 ns/op	134445098 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-8   	       8	 126689460 ns/op	134445070 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-8   	       8	 125782004 ns/op	134445106 B/op	 1700090 allocs/op
BenchmarkGetDomainStat-8   	       9	 125338795 ns/op	134445091 B/op	 1700090 allocs/op
PASS
ok  	github.com/fixme_my_friend/hw10_program_optimization	9.915s
```
После оптимизации
```
goos: linux
goarch: amd64
pkg: github.com/fixme_my_friend/hw10_program_optimization
cpu: Intel(R) Core(TM) i3-10100F CPU @ 3.60GHz
BenchmarkGetDomainStat-8   	  305187	      3837 ns/op	    7640 B/op	      28 allocs/op
BenchmarkGetDomainStat-8   	  305631	      3820 ns/op	    7640 B/op	      28 allocs/op
BenchmarkGetDomainStat-8   	  301830	      3806 ns/op	    7640 B/op	      28 allocs/op
BenchmarkGetDomainStat-8   	  300986	      3836 ns/op	    7640 B/op	      28 allocs/op
BenchmarkGetDomainStat-8   	  296218	      3824 ns/op	    7640 B/op	      28 allocs/op
PASS
ok  	github.com/fixme_my_friend/hw10_program_optimization	5.985s
```
benchstat
```
name             old time/op    new time/op    delta
GetDomainStat-8     126ms ± 1%       0ms ± 0%  -100.00%  (p=0.008 n=5+5)

name             old alloc/op   new alloc/op   delta
GetDomainStat-8     134MB ± 0%       0MB ± 0%   -99.99%  (p=0.008 n=5+5)

name             old allocs/op  new allocs/op  delta
GetDomainStat-8     1.70M ± 0%     0.00M ± 0%  -100.00%  (p=0.008 n=5+5)
```